condicionales y string

+++ TiposDeDatosNumericos/main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-    unsigned long  long var1 = 65535;
+    unsigned short var1 = 65535;
 
     cout << var1 << endl;
     var1 = var1 + 1;
